Conceptual, BatiTokenV2.
That was a conceptual contract to migrate my first token to ERC20 standard and i thought maybe i can improve it but i realized the main contract had a huge problem that i can't fix it anymore. So i gave up to rescue my own old token and the contract. Sadly, it is only a trash now. :(